requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Database & Cache
marshmallow>=3.0,<4.0
pymilvus>=2.4.0
redis==5.0.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
pgvector

# Machine Learning - RetinaFace & ArcFace (ONNX direct)
# insightface removed — fallback only, doesn't build on Python 3.13
opencv-python-headless>=4.10.0
numpy>=2.0.0
onnxruntime>=1.18.0

# Authentication & Security - JWT + bcrypt
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Pillow>=10.3.0
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
